backend/requirements.txt
Flask>=1.1.2
flask-bcrypt>=0.7.1
flask-cors>=3.0.9
flask-restx>=0.4.0
Flask-SQLAlchemy>=2.4.4
gunicorn>=20.0.4
psycopg2-binary>=2.8.6
flask-jwt-extended>=4.3.1
python-dateutil>=2.8.1
pytest>=7.1.2
ccxt>=1.81.14
pandas>=1.4.2
cryptocompare>=0.7.5
schedule>=1.1.0

